% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_distances.R
\name{find_hamming_dist}
\alias{find_hamming_dist}
\title{Find hamming distance}
\usage{
find_hamming_dist(query, reference, cols_to_match, ncores)
}
\arguments{
\item{query}{The query data frame}

\item{reference}{The reference data frame}

\item{cols_to_match}{The columns to match}

\item{ncores}{The number of cores to use}
}
\value{
A data frame with the hamming distance
}
\description{
Find hamming distance
}
