% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_distances.R
\name{find_levenshtein_dist}
\alias{find_levenshtein_dist}
\title{Find levenshtein distance}
\usage{
find_levenshtein_dist(query, reference, cols_to_match, ncores)
}
\arguments{
\item{query}{The query data frame}

\item{reference}{The reference data frame}

\item{cols_to_match}{The columns to match}

\item{ncores}{The number of cores to use}
}
\value{
A data frame with the levenshtein distance
}
\description{
Find levenshtein distance
}
