% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_distances.R
\name{find_min_distances}
\alias{find_min_distances}
\title{Find minimum distances}
\usage{
find_min_distances(output)
}
\arguments{
\item{output}{The output data frame}
}
\value{
A data frame with the minimum
}
\description{
Find minimum distances
}
