% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{get_reference}
\alias{get_reference}
\title{Get reference data frame}
\usage{
get_reference(antigen)
}
\arguments{
\item{antigen}{The antigen to get the reference data frame for}
}
\value{
A data frame with the reference data
}
\description{
Get reference data frame
}
