% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_cols.R
\name{lookup_cols}
\alias{lookup_cols}
\title{Lookup columns to match}
\usage{
lookup_cols(
  query,
  reference,
  heavyCDR3,
  heavyV = NA,
  heavyJ = NA,
  lightCDR3 = NA,
  lightV = NA,
  lightJ = NA
)
}
\arguments{
\item{query}{The query data frame}

\item{reference}{The reference data frame}

\item{heavyCDR3}{The heavy chain CDR3 sequence}

\item{heavyV}{The heavy chain V gene}

\item{heavyJ}{The heavy chain J gene}

\item{lightCDR3}{The light chain CDR3 sequence}

\item{lightV}{The light chain V gene}

\item{lightJ}{The light chain J gene}
}
\value{
A list of columns to match
}
\description{
Lookup columns to match
}
