% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preflight_checks.R
\name{preflight_checks}
\alias{preflight_checks}
\title{Preflight checks}
\usage{
preflight_checks(ncores, dist_method, output_dir)
}
\arguments{
\item{ncores}{The number of cores to use}

\item{dist_method}{The distance method to use}

\item{output_dir}{The output directory}
}
\description{
Preflight checks
}
