% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preflight_checks.R
\name{preflight_query}
\alias{preflight_query}
\title{Query checks}
\usage{
preflight_query(query, cols_to_match)
}
\arguments{
\item{query}{The query data frame}

\item{cols_to_match}{The columns to match}
}
\value{
A data frame with the preflight checks passed
}
\description{
Query checks
}
