% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preflight_checks.R
\name{preflight_reference}
\alias{preflight_reference}
\title{Reference checks}
\usage{
preflight_reference(reference, cols_to_match)
}
\arguments{
\item{reference}{The reference data frame}

\item{cols_to_match}{The columns to match}
}
\value{
A data frame with the preflight check passed
}
\description{
Reference checks
}
