% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_distances.R
\name{find_min_distances}
\alias{find_min_distances}
\title{Find minimum distances}
\usage{
find_min_distances(output, cols_to_match, output_dir = NULL)
}
\arguments{
\item{output}{The output data frame}

\item{cols_to_match}{The columns to match}

\item{output_dir}{The output directory}
}
\value{
A data frame with the minimum
}
\description{
Find minimum distances
}
